ADD: Shipping function init

+++ src/main/java/com/michael/shipping_system/model/RequestOrderCreate.java
@@ -7,9 +7,9 @@ import lombok.Setter;
 @Setter
 public class RequestOrderCreate {
 
-    private Long sendUserId;
+    private Integer sendUserId;
 
-    private Long receiptUserId;
+    private Integer receiptUserId;
 
     private Integer pickupLocationId;
